+++ packages/docusaurus/src/ssg/ssgTemplate.html.ts
